Modified the LSU
Regarding our meeting, it was mentioned that in case of misalignment, specially in the load case, the `lsu` requests the nearest lower and upper aligned addresses and then manipulate and mask them together to get the requested misaligned data. It was said that the driver preserve the data as if the two addresses contain the same data content and therefore, the scoreboard deals with one sequence item and rotate the data as needed.

It was not the generalized case for checking. Thus, I modified:

1- `lsu_driver`: to change the data in the next address.

2- `lsu_scoreboard`: to get two sequence items and manipulates them in the same manner as the `lsu` does.

3- `lsu_sequence_item`: Added another data bus (`data_rdata_next_i`) as it represents the data in the following address for the misalignment case. Also, I added `do_copy` function to be used in the `lsu_scoreboard` for saving the first transaction before retrieving the second transaction from the `tlm_fifo`.